修改数据类型

+++ tj-user/src/main/java/com/tianji/user/service/impl/UserServiceImpl.java
@@ -126,7 +126,7 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
         // 1.验证码校验
         codeService.verifyCode(user.getCellPhone(), code);
         // 2.判断手机号是否存在
-        Integer count = lambdaQuery().eq(User::getCellPhone, user.getCellPhone()).count();
+        Integer count = Math.toIntExact(lambdaQuery().eq(User::getCellPhone, user.getCellPhone()).count());
         if (count > 0) {
             throw new BadRequestException(PHONE_ALREADY_EXISTS);
         }
